requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0

# Authentication & Authorization
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
python-social-auth[django]==0.3.6
social-auth-app-django==5.4.0

# Monitoring & Logging
sentry-sdk[django]==1.38.0
django-extensions==3.2.3

# Communication
twilio==8.10.1
sendgrid==6.10.0

# Health checks & Monitoring
django-health-check==3.17.0
django-prometheus==2.3.1

# API Documentation
drf-spectacular==0.27.0

# Development
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2